refactor: Standardize code formatting in GraphQL types for improved readability

+++ src/main/java/org/amalitech/bloggingplatformspring/graphql/resolvers/GraphQLQueryResolver.java
@@ -63,7 +63,7 @@ public class GraphQLQueryResolver {
                 .collect(Collectors.toList()));
         graphQlPostPage.setPageNumber(posts.page());
         graphQlPostPage.setPageSize(posts.size());
-        graphQlPostPage.setTotalElements((long) posts.totalElements());
+        graphQlPostPage.setTotalElements(posts.totalElements());
         graphQlPostPage.setTotalPages((int) Math.ceil((double) posts.totalElements() / posts.size()));
 
         return graphQlPostPage;
